+++ crates/explorer/src/lib.rs
@@ -1,6 +1,0 @@
-//! The revive explorer leverages debug info to get insights into emitted code.
-
-pub mod dwarfdump;
-pub mod dwarfdump_analyzer;
-pub mod location_mapper;
-pub mod yul_phaser;